backend/requirements.txt
fastapi
uvicorn
pymongo
pydantic
motor
python-dotenv
python-jose[cryptography]
PyJWT
cryptography
requests
email-validator
httpx
python-multipart
aioboto3
Pillow
pillow-heif
